Added room, person, thing, instead rules
I added a Yoga Instructor who is a person, but not someone who is
important in the story (Will add a more important person).  Also added
things like a Restroom door, which is locked, that leads to the
Bathroom.  Added a gun that unlocks door, but not anywhere in game yet.
 Also added some instead rules.