Added more Stuff and a puzzle
I added a security room and a locked door to get into the security
room.  You can find the key to the door in the Cafe.  The Cafe is
completely empty with nothing in it except for the key (will add things
into it later).  When you get into the security room, I added a red
switch which toggles the security camera in the gun store.  If the
switch is switched on, you can steal a gun without getting caught.
There is also a key in the security room that unlocks the cabinet,
which I just added, that hold a gun.